#13 MOVE TestScraperBloomberg to own file
